requirements.txt
nba_api>=1.4.1
pandas>=2.1.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
scikit-learn>=1.4.0
xgboost>=2.0.0
pyarrow>=15.0.0
pytest>=8.0.0
streamlit>=1.40.0
requests>=2.31.0
optuna>=3.6.0
